chore: add boleto tab
