+++ packages/pilot/src/containers/Settings/Company/style.css
@@ -1,0 +1,3 @@
+.tabsContainer {
+  margin-top: 24px;
+}
